+++ src/main/java/com/flashcard/repository/UserSeenCardRepository.java
@@ -52,5 +52,5 @@ public interface UserSeenCardRepository extends JpaRepository<UserSeenCard, Long
 
     int countByUserAndCardFlashcardTopicLessonYks(User user, YKS yks);
 
-    long countByUserAndCardFlashcardTopic(User user, Topic topic);
+    Integer countByUserAndCardFlashcardTopic(User user, Topic topic);
 }
